docs: sendpay timestamps are u32
sendpay serializes the payment with json_add_payment_fields(),
which emits both timestamps with json_add_u32 (struct
wallet_payment keeps them as u32):

  lightningd/pay.c:158,160
    json_add_u32(response, "created_at", t->timestamp);
    json_add_u32(response, "completed_at", *t->completed_at);
  wallet/wallet.h:363-364

Changelog-None

+++ cln-rpc/src/model.rs
@@ -5870,7 +5870,7 @@ pub mod responses {
 	    #[serde(skip_serializing_if = "Option::is_none")]
 	    pub bolt12: Option<String>,
 	    #[serde(skip_serializing_if = "Option::is_none")]
-	    pub completed_at: Option<u64>,
+	    pub completed_at: Option<u32>,
 	    #[serde(skip_serializing_if = "Option::is_none")]
 	    pub destination: Option<PublicKey>,
 	    #[serde(skip_serializing_if = "Option::is_none")]
@@ -5886,7 +5886,7 @@ pub mod responses {
 	    // Path `SendPay.status`
 	    pub status: SendpayStatus,
 	    pub amount_sent_msat: Amount,
-	    pub created_at: u64,
+	    pub created_at: u32,
 	    pub created_index: u64,
 	    pub groupid: u64,
 	    pub id: u64,
